# Conflicts:
#   Laboratory_Management/Backend/LabolaryManagement/TestOrder.Application/Services/Booking/BookingService.cs
#   Laboratory_Management/Backend/LabolaryManagement/TestOrder.Presentation/appsettings.json
#   database/TestOrderDBScript.sql

+++ Laboratory_Management/Backend/LabolaryManagement/Instrument.Domain/Enums/RunStatus.cs
@@ -3,7 +3,7 @@
 /// <summary>
 /// Trạng thái của một lần chạy xét nghiệm (run)
 /// </summary>
-public enum RunStatus : byte
+public enum RunStatus : int
 {
     /// <summary>
     /// Đang chạy
